podmanv2 container inspect
add ability to inspect a container
