+++ cmd/podmanV2/containers/inspect.go
@@ -1,8 +1,15 @@
 package containers
 
 import (
+	"context"
+	"fmt"
+	"os"
+	"strings"
+	"text/template"
+
 	"github.com/containers/libpod/cmd/podmanV2/registry"
 	"github.com/containers/libpod/pkg/domain/entities"
+	jsoniter "github.com/json-iterator/go"
 	"github.com/spf13/cobra"
 )
 
@@ -12,11 +19,15 @@ var (
 		Use:     "inspect [flags] CONTAINER",
 		Short:   "Display the configuration of a container",
 		Long:    `Displays the low-level information on a container identified by name or ID.`,
-		PreRunE: inspectPreRunE,
+		PreRunE: preRunE,
 		RunE:    inspect,
 		Example: `podman container inspect myCtr
   podman container inspect -l --format '{{.Id}} {{.Config.Labels}}'`,
 	}
+)
+
+var (
+	inspectOptions entities.ContainerInspectOptions
 )
 
 func init() {
@@ -25,18 +36,40 @@ func init() {
 		Command: inspectCmd,
 		Parent:  containerCmd,
 	})
-}
-
-func inspectPreRunE(cmd *cobra.Command, args []string) (err error) {
-	err = preRunE(cmd, args)
-	if err != nil {
-		return
+	flags := inspectCmd.Flags()
+	flags.StringVarP(&inspectOptions.Format, "format", "f", "", "Change the output format to a Go template")
+	flags.BoolVarP(&inspectOptions.Latest, "latest", "l", false, "Act on the latest container podman is aware of")
+	flags.BoolVarP(&inspectOptions.Size, "size", "s", false, "Display total file size")
+	if registry.IsRemote() {
+		_ = flags.MarkHidden("latest")
 	}
-
-	_, err = registry.NewImageEngine(cmd, args)
-	return err
 }
 
 func inspect(cmd *cobra.Command, args []string) error {
+	responses, err := registry.ContainerEngine().ContainerInspect(context.Background(), args, inspectOptions)
+	if err != nil {
+		return err
+	}
+	if inspectOptions.Format == "" {
+		b, err := jsoniter.MarshalIndent(responses, "", "  ")
+		if err != nil {
+			return err
+		}
+		fmt.Println(string(b))
+		return nil
+	}
+	format := inspectOptions.Format
+	if !strings.HasSuffix(format, "\n") {
+		format += "\n"
+	}
+	tmpl, err := template.New("inspect").Parse(format)
+	if err != nil {
+		return err
+	}
+	for _, i := range responses {
+		if err := tmpl.Execute(os.Stdout, i); err != nil {
+			return err
+		}
+	}
 	return nil
 }
