+++ pkg/domain/entities/containers.go
@@ -81,3 +81,13 @@ type RmReport struct {
 	Err error
 	Id  string
 }
+
+type ContainerInspectOptions struct {
+	Format string
+	Latest bool
+	Size   bool
+}
+
+type ContainerInspectReport struct {
+	*define.InspectContainerData
+}
